+++ templates/wc-theme/functions.php
@@ -29,9 +29,9 @@ function my_enqueue_scripts() {
   wp_enqueue_style( 'dashicons', get_stylesheet_uri(), 'dashicons' ); // WP native icons
 
   // JavaScript
-  wp_enqueue_script( 'my-fastclick', $js_dir . '/vendor/fastclick.min.js', false, false, true );
-  wp_enqueue_script( 'my-magnific', $js_dir . '/vendor/magnific.min.js', array('jquery'), false, true );
-  wp_enqueue_script( 'my-slick', $js_dir . '/vendor/slick.min.js', array('jquery'), false, true );
+  wp_enqueue_script( 'my-fastclick', $js_dir . '-vendor/fastclick.min.js', false, false, true );
+  wp_enqueue_script( 'my-magnific', $js_dir . '-vendor/magnific.min.js', array('jquery'), false, true );
+  wp_enqueue_script( 'my-slick', $js_dir . '-vendor/slick.min.js', array('jquery'), false, true );
 
   wp_enqueue_script( 'my-app', $js_dir . '/app.js', array('jquery'), false, true );
   wp_enqueue_script( 'my-app-wc', $js_dir . '/app-wc.js', array('jquery'), false, true );